+++ apps/app/src/views/components/Auth.tsx
@@ -1,5 +1,4 @@
 import { Button, Stack } from '@chakra-ui/react';
-import { CopyIcon } from '@chakra-ui/icons';
 
 import { usePhoton } from '@photonhealth/react';
 import { getSettings } from '@client/settings';
@@ -13,33 +12,14 @@ interface AuthProps {
 
 export const Auth = (props: AuthProps) => {
   const { returnTo } = props;
-  const { user, isLoading, isAuthenticated, getToken, login, logout } = usePhoton();
+  const { user, isLoading, isAuthenticated, login, logout } = usePhoton();
   const orgSettings = user?.org_id in settings ? settings[user?.org_id] : settings.default;
-
-  const getAccessToken = async () => {
-    try {
-      const token = await getToken();
-      navigator.clipboard.writeText(token);
-    } catch (e) {
-      console.error(e);
-    }
-  };
 
   if (isLoading) return <Button isLoading loadingText="Loading" colorScheme="gray" />;
 
   if (isAuthenticated)
     return (
       <Stack direction="row" spacing={4}>
-        <Button
-          borderColor="blue.500"
-          textColor="blue.500"
-          colorScheme="blue"
-          rightIcon={<CopyIcon />}
-          variant="outline"
-          onClick={getAccessToken}
-        >
-          Auth Token
-        </Button>
         <Button
           colorScheme="brand"
           onClick={() => {
